+++ src/components/FeedbackForm/FeedbackForm.tsx
@@ -1,4 +1,4 @@
-import { FormControl, FormHelperText, FormLabel, Textarea, Button, HStack, useToast, Center } from '@chakra-ui/react';
+import { Button, Center, FormControl, FormHelperText, FormLabel, HStack, Textarea, useToast } from '@chakra-ui/react';
 import { useSession } from 'next-auth/react';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
